Working Starter Script
First fix is to invoke charon via the starter application or else the
the ipsec.conf file doesn't get imported.  Following that it was observed
that packets traversing from the node port to the container undergo SNAT
and we see that as the right address.  This can be dynamically filled in
from the default route.