Working Site-to-site Configuration
The previous configuration worked as an IPSEC responder but when acting
as an initiator it lacked any way of addressing the remote end.  To fix this
we now support the remote L3 address via configuration, the correct IPSEC
connection is selected by default neither source or remote IP address are
interrogated and we just default to %any behavior.

This however still has one issue in that as the gateway's ingress and egress ports
are the same we see ICMP redirects which may install routes and bypass the
IPSEC transform entirely.  Opened #1 to address, however support is limited and
not production ready.